json error bug fixed

+++ get_classLinks.py
@@ -15,18 +15,18 @@ date = datetime.date.today()
 if os.path.exists(f"datadump{Previous_Date}.json"):
     os.remove(f"datadump{Previous_Date}.json")
 
-sessionID = config.sessionID
-GYOC = {"ASP.NET_SessionId": sessionID}
-data = {"text": date}
-url = "https://glauniversity.in:8085/MyAccount/DutyDetails"
-r = requests.post(url, data=data, cookies=GYOC)
-
-try:
-    with open(f"datadump{date}.json", 'w') as f:
-        f.write(r.text)
-except json.decoder.JSONDecodeError as e:
-    print("SessionID has expired, will need a new one in order to continue")
-    exit()
+if not os.path.exists(f"datadump{date}.json"):
+    sessionID = config.sessionID
+    GYOC = {"ASP.NET_SessionId": sessionID}
+    data = {"text": date}
+    url = "https://glauniversity.in:8085/MyAccount/DutyDetails"
+    r = requests.post(url, data=data, cookies=GYOC)
+    try:
+        with open(f"datadump{date}.json", 'w') as f:
+            f.write(r.text)
+    except json.decoder.JSONDecodeError as e:
+        print("SessionID has expired, will need a new one in order to continue")
+        exit()
 with open(f"datadump{date}.json") as f:
     y = json.load(f)
 # For distribution
